% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{choropleth}
\alias{choropleth}
\title{Calculates an isochrone map from a single origin}
\usage{
choropleth(output.dir, otpcon, originPoints, originPolygons, destinationPoints,
  destinationPointsRow = 1, startDateAndTime = "2018-08-18 12:00:00",
  modes = "WALK, TRANSIT", maxWalkDistance = 1000, walkReluctance = 2,
  walkSpeed = 1.5, bikeSpeed = 5, minTransferTime = 1, maxTransfers = 5,
  wheelchair = F, arriveBy = F, durationCutoff = 60, waitingCutoff = 10,
  transferCutoff = 1, palColor = "Blues", palColorCat = c("#820e0e",
  "#407746"), mapZoom = 12)
}
\arguments{
\item{otpcon}{OTP router URL}

\item{originPoints}{The variable containing origin(s), see ?importLocationData}

\item{originPolygons}{The variable containing origin(s) polygon(s), see ?importGeojsonData}

\item{destinationPoints}{The variable containing destination(s), see ?importLocationData}

\item{destinationPointsRow}{The row of destinationPoints to be used, defaults to 1}

\item{startDateAndTime}{in 'YYYY-MM-DD HH:MM:SS' format}

\item{modes}{defaults to 'TRANSIT, WALK'}

\item{maxWalkDistance}{in meters, defaults to 1000}

\item{walkReluctance}{defaults to 2 (range 0 - 20)}

\item{walkSpeed}{in m/s, defaults to 1.4}

\item{bikeSpeed}{in m/s, defaults to 4.3}

\item{minTransferTime}{in minutes, defaults to 1}

\item{maxTransfers}{defaults to 10}

\item{wheelchair}{defaults to FALSE}

\item{arriveBy}{defaults to FALSE}

\item{durationCutoff}{in minutes, defaults to 60}

\item{waitingCutoff}{in minutes, defaults to 10}

\item{transferCutoff}{defaults to 1}

\item{palColor}{the color palette of the map, defaults to 'Blues'}

\item{palColorCat}{the color palette of the catergorical map, defaults to c("#820e0e", "#407746")}

\item{mapZoom}{defaults to 12}
}
\value{
Returns a number of maps (duration, wait time, transfers) to the output directory
}
\description{
.. content for \details{} ..
}
\examples{
  choropleth(
    output.dir = 'C:\\Users\\User\\Documents',
    otpcon,
    originPoints,
    originPolygons,
    destinationPoints,
    startdDateAndTime = "2018-08-18 12:00:00"
  )
}
\author{
Michael Hodge
}
