% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs.R
\name{cleanGTFS}
\alias{cleanGTFS}
\title{Cleans GTFS feed data}
\usage{
cleanGTFS(gtfs.dir, gtfs.filename)
}
\arguments{
\item{gtfs.dir}{The directory for the GTFS files}

\item{gtfs.filename}{The name of the GTFS files}

\item{train}{defaults to TRUE}
}
\value{
Returns a cleaned GTFS zip file
}
\description{
Generates a new, cleaned GTFS zip file.
}
\author{
Michael Hodge
}
