% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{importGeojsonData}
\alias{importGeojsonData}
\title{Imports Geojson Data}
\usage{
importGeojsonData(src)
}
\arguments{
\item{src}{.Geojson file source}
}
\value{
Large SpatialPolygonsDataFrame of origin polygons
}
\description{
Imports polygon data (origin) from a geojson file
}
\examples{
originPolygons <- importGeojsonData('C:\\Users\\User\\Documents\\origins.geojson')
}
\author{
Michael Hodge
}
