% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{importLocationData}
\alias{importLocationData}
\title{Imports Location Data}
\usage{
importLocationData(src)
}
\arguments{
\item{src}{.CSV file source}
}
\value{
R dataframe of origin or destination points
}
\description{
Imports location data (origin or destination) from a comma separated file
}
\examples{
originPoints <- importLocationData('C:\\Users\\User\\Documents\\origins.csv')
destinationPoints <- importLocationData('C:\\Users\\User\\Documents\\destinations.csv')
}
\author{
Michael Hodge
}
