% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isochrone.R
\name{isochrone}
\alias{isochrone}
\title{Generates an isochrone map from a single origin}
\usage{
isochrone(output.dir, otpcon, originPoints, originPointsRow = 1,
  destinationPoints, destinationPointsRow = 1,
  startDateAndTime = "2018-08-18 12:00:00", modes = "WALK, TRANSIT",
  maxWalkDistance = 1000, walkReluctance = 2, walkSpeed = 1.5,
  bikeSpeed = 5, minTransferTime = 1, maxTransfers = 5,
  wheelchair = F, arriveBy = F, isochroneCutOffs = c(30, 60, 90),
  map = FALSE, geojson = FALSE, palColor = "Blues", mapZoom = 12)
}
\arguments{
\item{output.dir}{The directory for the output files}

\item{otpcon}{OTP router URL}

\item{originPoints}{The variable containing origin(s), see ?importLocationData}

\item{originPointsRow}{The row of originPoints to be used, defaults to 1}

\item{destinationPoints}{The variable containing destination(s) see ?importLocationData}

\item{startDateAndTime}{in 'YYYY-MM-DD HH:MM:SS' format}

\item{modes}{defaults to 'TRANSIT, WALK'}

\item{maxWalkDistance}{in meters, defaults to 1000}

\item{walkReluctance}{defaults to 2 (range 0 - 20)}

\item{walkSpeed}{in m/s, defaults to 1.4}

\item{bikeSpeed}{in m/s, defaults to 4.3}

\item{minTransferTime}{in minutes, defaults to 1}

\item{maxTransfers}{defaults to 10}

\item{wheelchair}{defaults to FALSE}

\item{arriveBy}{defaults to FALSE}

\item{isochroneCutOffs}{a list of cutoffs in minutes, defaults to c(30, 60, 90)}

\item{map}{specify whether you want to output a map}

\item{geojson}{specific whether you want to output a GeoJSON file, defaults to FALSE}

\item{palColor}{the color palette of the map, defaults to 'Blues'}

\item{mapZoom}{defaults to 12}
}
\value{
Saves map as a png and journey details as CSV to output directory
}
\description{
Generates an isochrone map from a single origin and checks whether destinations
fall within isochrone, and if so, at what cutoff time amount.
}
\examples{
  isochrone(
    output.dir = 'C:\\Users\\User\\Documents',
    otpcon,
    originPoints,
    destinationPoints,
    startDateAndTime = "2018-08-18 12:00:00"
  )
}
\author{
Michael Hodge
}
