% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isochrone-multi-intersect.R
\name{isochroneMultiIntersect}
\alias{isochroneMultiIntersect}
\title{Generates a GeoJSON of intersecting polygons for multiple origins}
\usage{
isochroneMultiIntersect(output.dir, otpcon, originPoints,
  destinationPoints, startDateAndTime = "2018-08-18 12:00:00",
  modes = "WALK, TRANSIT", maxWalkDistance = 1000,
  walkReluctance = 2, walkSpeed = 1.5, bikeSpeed = 5,
  minTransferTime = 1, maxTransfers = 5, wheelchair = F,
  arriveBy = F, isochroneCutOffs = 60, mapOutput = FALSE,
  geojsonOutput = TRUE, mapMarkerColours = "Greys",
  mapPolygonColours = "#6BAED6", mapZoom = 12,
  mapPolygonLineWeight = 5, mapPolygonLineOpacity = 1,
  mapPolygonFillOpacity = 0.6, mapMarkerOpacity = 1,
  mapLegendOpacity = 0.5)
}
\arguments{
\item{output.dir}{The directory for the output files}

\item{otpcon}{The OTP router URL}

\item{originPoints}{The variable containing origin(s), see ?importLocationData}

\item{destinationPoints}{The variable containing destination(s) see ?importLocationData}

\item{startDateAndTime}{The start time and date, in 'YYYY-MM-DD HH:MM:SS' format}

\item{modes}{The mode of the journey, defaults to 'TRANSIT, WALK'}

\item{maxWalkDistance}{The maximum walking distance, in meters, defaults to 1000 m}

\item{walkReluctance}{The reluctance of walking-based routes, defaults to 2 (range 0 (lowest) - 20 (highest))}

\item{walkSpeed}{The walking soeed, in meters per second, defaults to 1.4 m/s}

\item{bikeSpeed}{The cycling speed, in meters per second, defaults to 4.3 m/s}

\item{minTransferTime}{The maximum transfer time, in minutes, defaults to 0 mins (no time specified)}

\item{maxTransfers}{The maximum number of transfers, defaults to 10}

\item{wheelchair}{If TRUE, uses on wheeelchair friendly stops, defaults to FALSE}

\item{arriveBy}{Selects whether journey starts at startDateandTime (FALSE) or finishes (TRUE), defaults to FALSE}

\item{isochroneCutOffs}{Provide a list of cutoffs in minutes, defaults to c(30, 60, 90)}

\item{mapOutput}{Specifies whether you want to output a map, defaults to FALSE}

\item{geojsonOutput}{Specifies whether you want to output a GeoJSON file, defaults to FALSE}

\item{mapMarkerColours}{the color palette of the markers, defaults to 'Greys'}

\item{mapPolygonColours}{the color palette of the poygon, defaults to 'Blue'}

\item{mapZoom}{The zoom level of the map, defaults to 12}

\item{mapPolygonLineWeight}{Specifies the weight of the polygon, defaults to 5 px}

\item{mapPolygonLineOpacity}{Specifies the opacity of the polygon line, defaults to 1 (solid)}

\item{mapPolygonFillOpacity}{Specifies the opacity of the polygon fill, defaults to 0.6}

\item{mapMarkerOpacity}{Specifies the opacity of the marker, defaults to 1 (solid)}

\item{mapLegendOpacity}{Specifies the opacity of the legend, defaults to 0.5}
}
\value{
Saves map as a Geojson of the intersection area to output directory (optional: a map in PNG and HTML formats)
}
\description{
Calculates an polygon for the intersection between multiple origins and finds the intersection.
Saves polygon as a GeoJSON file.
}
\examples{
  isochroneMultiIntersect(
    output.dir = 'C:\\Users\\User\\Documents',
    otpcon,
    originPoints,
    destinationPoints,
    startDateAndTime = "2018-08-18 12:00:00"
  )
}
\author{
Michael Hodge
}
