% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location-validator.R
\name{locationValidatorIsochrone}
\alias{locationValidatorIsochrone}
\title{Checks the validity of locations for isochrones}
\usage{
locationValidatorIsochrone(output.dir, otpcon = otpcon,
  locationPoints = originPoints, modes = "WALK", cutoff = 20,
  infoPrint = T)
}
\arguments{
\item{output.dir}{The directory for the output files}

\item{otpcon}{The OTP router URL, see ?otpcon for details}

\item{locationPoints}{The variable containing origin(s), see ?importLocationData for details}

\item{modes}{The mode of the journey, defaults to 'WALK'}

\item{cutoff}{Specify the isochrone cutoff distance, defaults to 20 mins}

\item{infoPrint}{Specifies whether you want some information printed to the console or not, default is TRUE}
}
\description{
Calculates the journey time and details between multiple origins and/or destinations. 
A comma separated value file of journey details is saved in the output folder.
}
\author{
Michael Hodge
}
