% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nominatimNodeSearch}
\alias{nominatimNodeSearch}
\title{Finds list of nodes near a lat and lon}
\usage{
nominatimNodeSearch(lat, lon)
}
\arguments{
\item{lat}{latitude}

\item{lon}{longitude}
}
\value{
dataframe of lats and lons of nearest nodes
}
\description{
Finds list of nodes near a lat and lon
}
\author{
Michael Hodge
}
