% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nominatimObjectSearch}
\alias{nominatimObjectSearch}
\title{Finds lthe nearest object to a lat and lon}
\usage{
nominatimObjectSearch(lat, lon, object)
}
\arguments{
\item{lat}{latitude}

\item{lon}{longitude}

\item{object}{the object you want to search for (e.g. pub)}
}
\value{
dataframe of lats and lons of nearest nodes
}
\description{
Finds lthe nearest object to a lat and lon
}
\author{
Michael Hodge
}
