% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp.R
\name{otpChoropleth}
\alias{otpChoropleth}
\title{Returns trip details for choropleth analysis}
\usage{
otpChoropleth(otpcon, from, to, modes, detail = TRUE, date, time,
  maxWalkDistance = 800, walkReluctance = 2, arriveBy = FALSE,
  transferPenalty = 0, minTransferTime = 1, walkSpeed = 1.4,
  bikeSpeed = 4.3, maxTransfers = 10, wheelchair = FALSE)
}
\arguments{
\item{otpcon}{OTP router URL}

\item{from}{'lat, lon' decimal degrees}

\item{to}{'lat, lon' decimal degrees}

\item{modes}{defaults to 'TRANSIT, WALK'}

\item{detail}{defaults to TRUE}

\item{date}{'YYYY-MM-DD' format}

\item{time}{12 hour format}

\item{maxWalkDistance}{in meters, defaults to 800}

\item{walkReluctance}{defaults to 2 (range 0 - 20)}

\item{arriveBy}{defaults to FALSE}

\item{transferPenalty}{defaults to 0}

\item{minTransferTime}{in minutes, defaults to 1}

\item{walkSpeed}{in m/s, defaults to 1.4}

\item{bikeSpeed}{in m/s, defaults to 4.3}

\item{maxTransfers}{defaults to 10}

\item{wheelchair}{defaults to FALSE}
}
\value{
A list comprising status and journey itineraries (duration, walk time, transit time, waiting time and transfers)
}
\description{
A light version of the OtpTripTime function designed to work for the choropleth
analysis
}
\author{
Michael Hodge
}
