% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp.R
\name{otpConnect}
\alias{otpConnect}
\title{Get OTP connection URL}
\usage{
otpConnect(hostname = "localhost", router = "default", port = "8080",
  ssl = FALSE)
}
\arguments{
\item{hostname}{defaults to 'localhost'}

\item{router}{defaults to 'default'}

\item{port}{defaults to 8080}

\item{ssl}{defaults to FALSE}
}
\value{
The OTP server URL as https://hostname:port/otp/routers/router or http://hostname:port/otp/routers/router
}
\description{
Connects to the OpenTripPlanner (OTP) server URL
}
\author{
Michael Hodge
}
