% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp.R
\name{otpIsochrone}
\alias{otpIsochrone}
\title{Return isochrone from OTP}
\usage{
otpIsochrone(otpcon, from, to, modes, cutoff = c(30, 60, 90),
  batch = TRUE, date, time, maxWalkDistance = 800,
  walkReluctance = 2, walkSpeed = 1.4, bikeSpeed = 4.3,
  minTransferTime = 1, maxTransfers = 10, wheelchair = FALSE,
  arriveBy = FALSE)
}
\arguments{
\item{otpcon}{OTP router URL}

\item{from}{'lat, lon' decimal degrees}

\item{modes}{defaults to 'TRANSIT, WALK'}

\item{cutoff}{list in minutes, defaults to c(30,60,90)}

\item{batch}{defaults to TRUE}

\item{date}{'YYYY-MM-DD' format}

\item{time}{12 hour format}

\item{maxWalkDistance}{in meters, defaults to 800}

\item{walkReluctance}{defaults to 2 (range 0 - 20)}

\item{walkSpeed}{in m/s, defaults to 1.4}

\item{bikeSpeed}{in m/s, defaults to 4.3}

\item{minTransferTime}{in minutes, defaults to 1}

\item{maxTransfers}{defaults to 10}

\item{wheelchair}{defaults to FALSE}

\item{arriveBy}{defaults to FALSE}
}
\value{
A list comprising status and isochrones as a text object
}
\description{
Returns an isochrone from OTP for a specified maximum duration and cutoff intervals based on the origin(s) location(s)
}
\author{
Michael Hodge
}
