% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp.R
\name{otpJourneyChecker}
\alias{otpJourneyChecker}
\title{Returns is journey is possible from OTP}
\usage{
otpJourneyChecker(otpcon, from_name, from_lat_lon, to_name, to_lat_lon,
  modes = "WALK", date, time, maxWalkDistance = 800,
  walkReluctance = 2, arriveBy = "false", transferPenalty = 0,
  minTransferTime = 1, walkSpeed = 1.4, bikeSpeed = 4.3,
  maxTransfers = 10, wheelchair = FALSE)
}
\arguments{
\item{otpcon}{OTP router URL}

\item{modes}{defaults to 'TRANSIT, WALK'}

\item{date}{'YYYY-MM-DD' format}

\item{time}{12 hour format}

\item{maxWalkDistance}{in meters, defaults to 800}

\item{walkReluctance}{defaults to 2 (range 0 - 20)}

\item{arriveBy}{defaults to FALSE}

\item{transferPenalty}{defaults to 0}

\item{minTransferTime}{in minutes, defaults to 1}

\item{walkSpeed}{in m/s, defaults to 1.4}

\item{bikeSpeed}{in m/s, defaults to 4.3}

\item{maxTransfers}{defaults to 10}

\item{wheelchair}{defaults to FALSE}

\item{from}{'lat, lon' decimal degrees}

\item{to}{'lat, lon' decimal degrees}

\item{preWaitTime}{in minutes, defaults to 60}
}
\value{
A list comprising status and duration (in minutes) of detail is FALSE, or list comprising status, and itineraries, trip details
and polylines dataframes
}
\description{
Returns if journey is possible.
}
\author{
Michael Hodge
}
