% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp.R
\name{otpTripDistance}
\alias{otpTripDistance}
\title{Returns trip distance from OTP}
\usage{
otpTripDistance(otpcon, from, to, modes = "WALK,TRANSIT")
}
\arguments{
\item{otpcon}{OTP router URL}

\item{from}{'lat, lon' decimal degrees}

\item{to}{'lat, lon' decimal degrees}

\item{modes}{defaults to 'TRANSIT, WALK'}
}
\value{
A list comprising status and distance (in meters)
}
\description{
Returns trip distance. The parameters are: from, to and transport mode(s).
}
\author{
Michael Hodge
}
