% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp.R
\name{otpTripTime}
\alias{otpTripTime}
\title{Returns trip time from OTP}
\usage{
otpTripTime(otpcon, from_name, from_lat_lon, to_name, to_lat_lon,
  modes = "WALK,TRANSIT", detail = TRUE, date, time,
  maxWalkDistance = 800, walkReluctance = 2, arriveBy = "false",
  transferPenalty = 0, minTransferTime = 1, walkSpeed = 1.4,
  bikeSpeed = 4.3, maxTransfers = 10, wheelchair = FALSE,
  preWaitTime = 60)
}
\arguments{
\item{otpcon}{OTP router URL}

\item{modes}{defaults to 'TRANSIT, WALK'}

\item{detail}{defaults to TRUE}

\item{date}{'YYYY-MM-DD' format}

\item{time}{12 hour format}

\item{maxWalkDistance}{in meters, defaults to 800}

\item{walkReluctance}{defaults to 2 (range 0 - 20)}

\item{arriveBy}{defaults to FALSE}

\item{transferPenalty}{defaults to 0}

\item{minTransferTime}{in minutes, defaults to 1}

\item{walkSpeed}{in m/s, defaults to 1.4}

\item{bikeSpeed}{in m/s, defaults to 4.3}

\item{maxTransfers}{defaults to 10}

\item{wheelchair}{defaults to FALSE}

\item{preWaitTime}{in minutes, defaults to 60}

\item{from}{'lat, lon' decimal degrees}

\item{to}{'lat, lon' decimal degrees}
}
\value{
A list comprising status and duration (in minutes) of detail is FALSE, or list comprising status, and itineraries, trip details
and polylines dataframes
}
\description{
Returns trip time in simple or detailed form. The parameters from, to, modes,
date and time must be specified in the function call, other parameters have
defaults set and are optional in the call.
}
\author{
Michael Hodge
}
