% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point-to-point.R
\name{pointToPoint}
\alias{pointToPoint}
\title{Calculates the journey for a single origin and destination.}
\usage{
pointToPoint(output.dir, otpcon, originPoints, originPointsRow = 1,
  destinationPoints, destinationPointsRow = 1,
  startDateAndTime = "2018-08-18 12:00:00", modes = "WALK, TRANSIT",
  maxWalkDistance = 1000, walkReluctance = 2, walkSpeed = 1.4,
  bikeSpeed = 4.3, minTransferTime = 1, maxTransfers = 10,
  wheelchair = F, arriveBy = F, preWaitTime = 60,
  transportColours = list(TRANSIT = "#D3D3D3", WALK = "#87CEEB", BUS =
  "#DF4343", RAIL = "#43DF48", CAR = "#DFAB43", BICYCLE = "#000000"),
  mapZoom = 12)
}
\arguments{
\item{output.dir}{The directory for the output files}

\item{otpcon}{OTP router URL}

\item{originPoints}{The variable containing origin(s), see ?importLocationData}

\item{originPointsRow}{The row of originPoints to be used, defaults to 1}

\item{destinationPoints}{The variable containing destination(s) see ?importLocationData}

\item{destinationPointsRow}{The row of destinationPoints to be used, defaults to 1}

\item{startDateAndTime}{in 'YYYY-MM-DD HH:MM:SS' format}

\item{modes}{defaults to 'TRANSIT, WALK'}

\item{maxWalkDistance}{in meters, defaults to 1000}

\item{walkReluctance}{defaults to 2 (range 0 - 20)}

\item{walkSpeed}{in m/s, defaults to 1.4}

\item{bikeSpeed}{in m/s, defaults to 4.3}

\item{minTransferTime}{in minutes, defaults to 0}

\item{maxTransfers}{defaults to 10}

\item{wheelchair}{defaults to FALSE}

\item{arriveBy}{defaults to FALSE}

\item{preWaitTime}{in minutes, defaults to 60}

\item{transportColours}{A list defining the colours to assign to each mode of transport.}

\item{mapZoom}{defaults to 12}
}
\value{
Saves map as a png and journey details as CSV to output directory
}
\description{
Calculates the journey time and details between a single origin and destination. Outputs (and saves) a map of
the polyline between origin and destination, as well as a CSV file of journey details.
}
\examples{
  pointToPoint(
    output.dir = 'C:\\Users\\User\\Documents',
    otpcon,
    originPoints,
    destinationPoints,
    startdDateAndTime = "2018-08-18 12:00:00"
  )
}
\author{
Michael Hodge
}
