% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point-to-point-nearest.R
\name{pointToPointNearest}
\alias{pointToPointNearest}
\title{Calculates the journey for a number of origins and/or destinations.}
\usage{
pointToPointNearest(output.dir, otpcon, originPoints, destinationPoints,
  journeyReturn = F, startDateAndTime = "2018-08-13 09:00:00",
  modes = "WALK, TRANSIT", maxWalkDistance = 1000,
  walkReluctance = 2, walkSpeed = 1.5, bikeSpeed = 5,
  minTransferTime = 1, maxTransfers = 5, wheelchair = F,
  arriveBy = F, preWaitTime = 15, nearestNum = 1)
}
\arguments{
\item{output.dir}{The directory for the output files}

\item{otpcon}{The OTP router URL, see ?otpcon for details}

\item{originPoints}{The variable containing origin(s), see ?importLocationData for details}

\item{destinationPoints}{The variable containing destination(s) see ?importLocationData for details}

\item{journeyReturn}{Specifies whether the journey should be calculated as a return or not (default is FALSE)}

\item{startDateAndTime}{The start time and date, in 'YYYY-MM-DD HH:MM:SS' format}

\item{modes}{The mode of the journey, defaults to 'TRANSIT, WALK'}

\item{maxWalkDistance}{The maximum walking distance, in meters, defaults to 1000 m}

\item{walkReluctance}{The reluctance of walking-based routes, defaults to 2 (range 0 (lowest) - 20 (highest))}

\item{walkSpeed}{The walking soeed, in meters per second, defaults to 1.4 m/s}

\item{bikeSpeed}{The cycling speed, in meters per second, defaults to 4.3 m/s}

\item{minTransferTime}{The maximum transfer time, in minutes, defaults to 0 mins (no time specified)}

\item{maxTransfers}{The maximum number of transfers, defaults to 10}

\item{wheelchair}{If TRUE, uses on wheeelchair friendly stops, defaults to FALSE}

\item{arriveBy}{Selects whether journey starts at startDateandTime (FALSE) or finishes (TRUE), defaults to FALSE}

\item{preWaitTime}{The maximum waiting time before a journey cannot be found, in minutes, defaults to 15 mins}

\item{nearestNum}{Specify the number of nearest pairs to return (default is 1)}
}
\value{
Saves journey details as comma separated value file to output directory
}
\description{
Calculates the journey time and details between multiple origins and/or destinations. 
A comma separated value file of journey details is saved in the output folder.
}
\examples{
  pointToPointNearest(
    output.dir = 'C:\\Users\\User\\Documents',
    otpcon,
    originPoints,
    destinationPoints,
    journeyReturn = TRUE,
    startDateAndTime = "2018-08-18 12:00:00",
    nearestNum = 1,
  )
}
\author{
Michael Hodge
}
