% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point-to-point-time.R
\name{pointToPointTime}
\alias{pointToPointTime}
\title{Calculates the journey between for a single origin and destination between a
start and end time and date.}
\usage{
pointToPointTime(output.dir, otpcon, originPoints, originPointsRow = 1,
  destinationPoints, destinationPointsRow = 1,
  startDateAndTime = "2018-08-18 12:00:00",
  endDateAndTime = "2018-08-18 15:00:00", timeIncrease = 60,
  modes = "WALK, TRANSIT", maxWalkDistance = 1000,
  walkReluctance = 2, walkSpeed = 1.5, bikeSpeed = 5,
  minTransferTime = 1, maxTransfers = 5, wheelchair = F,
  arriveBy = F, preWaitTime = 60, mapOutput = FALSE,
  mapPolylineColours = list(TRANSIT = "#000000", WALK = "#A14296", BUS =
  "#48C1B1", RAIL = "#4D7BC5", CAR = "#8D4084", BICYCLE = "#4AA6C3"),
  mapZoom = 12, mapPolylineWeight = 5, mapPolylineOpacity = 1,
  mapMarkerOpacity = 1, mapLegendOpacity = 1)
}
\arguments{
\item{output.dir}{The directory for the output files}

\item{otpcon}{The OTP router URL}

\item{originPoints}{The variable containing origin(s), see ?importLocationData}

\item{originPointsRow}{The row of originPoints to be used, defaults to 1}

\item{destinationPoints}{The variable containing destination(s) see ?importLocationData}

\item{destinationPointsRow}{The row of destinationPoints to be used, defaults to 1}

\item{startDateAndTime}{The start time and date, in 'YYYY-MM-DD HH:MM:SS' format}

\item{endDateAndTime}{The end time and date, in 'YYYY-MM-DD HH:MM:SS' format}

\item{timeIncrease}{The time increase in minutes, default 60}

\item{modes}{The mode of the journey, defaults to 'TRANSIT, WALK'}

\item{maxWalkDistance}{The maximum walking distance, in meters, defaults to 1000 m}

\item{walkReluctance}{The reluctance of walking-based routes, defaults to 2 (range 0 (lowest) - 20 (highest))}

\item{walkSpeed}{The walking soeed, in meters per second, defaults to 1.4 m/s}

\item{bikeSpeed}{The cycling speed, in meters per second, defaults to 4.3 m/s}

\item{minTransferTime}{The maximum transfer time, in minutes, defaults to 0 mins (no time specified)}

\item{maxTransfers}{The maximum number of transfers, defaults to 10}

\item{wheelchair}{If TRUE, uses on wheeelchair friendly stops, defaults to FALSE}

\item{arriveBy}{Selects whether journey starts at startDateandTime (FALSE) or finishes (TRUE), defaults to FALSE}

\item{preWaitTime}{The maximum waiting time before a journey cannot be found, in minutes, defaults to 60 mins}

\item{mapOutput}{Specifies whether you want to output a map, defaults to FALSE}

\item{mapPolylineColours}{A list defining the colours to assign to each mode of transport.}

\item{mapZoom}{The zoom level of the map, defaults to 12}

\item{mapPolylineWeight}{Specifies the weight of the polyline, defaults to 5 px}

\item{mapPolylineOpacity}{Specifies the opacity of the polyline, defaults to 1 (solid)}

\item{mapMarkerOpacity}{Specifies the opacity of the marker, defaults to 1 (solid)}

\item{mapLegendOpacity}{Specifies the opacity of the legend, defaults to 1 (solid)}
}
\value{
Saves journey details as CSV to output directory (optional: an animated map in GIF format)
}
\description{
Calculates the journey time and details between a single origin and destination between a start
and end time and date. A CSV file of journey details is saved in the output folder.
An animated map of the journey can also be saved as a GIF file.
}
\examples{
  pointToPointTime(
    output.dir = 'C:\\Users\\User\\Documents',
    otpcon,
    originPoints,
    destinationPoints,
    startDateAndTime = "2018-08-18 12:00:00",
    endDateAndTime = "2018-08-18 13:00:00",
    timeIncrease = 60
  )
}
\author{
Michael Hodge
}
