% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{postcodeComplete}
\alias{postcodeComplete}
\title{Postcode autocomplete}
\usage{
postcodeComplete(postcode)
}
\arguments{
\item{postcode}{UK postcode}
}
\value{
The latitude and longitude in decimal degrees.
}
\description{
Generates a postcode from a partial postcode
}
\examples{
postcodeToDecimalDegrees_backup('NP10 8XG')
}
\author{
Michael Hodge
}
