% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{postcodeToDecimalDegrees_backup}
\alias{postcodeToDecimalDegrees_backup}
\title{A backup postcode to latitude and longitude converter.}
\usage{
postcodeToDecimalDegrees_backup(postcode)
}
\arguments{
\item{postcode}{UK postcode}
}
\value{
The latitude and longitude in decimal degrees.
}
\description{
Generates a latitude and longitude (in decimal degrees) based on the postcode.
}
\author{
Michael Hodge
}
