% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.plots.exists.R
\name{qc.plots.exists}
\alias{qc.plots.exists}
\title{Plots Folder}
\usage{
qc.plots.exists(folder, stop = TRUE, quiet = FALSE)
}
\arguments{
\item{folder}{assessment folder.}

\item{stop}{whether to stop if test fails.}

\item{quiet}{whether to suppress messages.}
}
\value{
\code{TRUE} if test succeeds, otherwise an error message
(if \code{stop = TRUE}) or \code{FALSE} and a warning message
(if \code{stop = FALSE}).
}
\description{
Assert that a \verb{plots} folder exists.
}
\examples{
\dontrun{
qc.plots.scripts("z:/yft/2023/plots")
}

}
