% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs.R
\name{stopFIX}
\alias{stopFIX}
\title{Fixes 0,0 in stops GTFS file}
\usage{
stopFIX(gtfs.dir, stops.filename)
}
\arguments{
\item{gtfs.dir}{The directory for the stop file}

\item{stops.filename}{The name of the stop file}
}
\value{
Returns a stop file with fixed lats and lons
}
\description{
Fixes 0,0 in stops GTFS file
}
\author{
Michael Hodge
}
