% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{choropleth}
\alias{choropleth}
\title{Generates a series of choropleth maps using multiple origins to a single destination.}
\usage{
choropleth(output.dir, otpcon, originPoints, originPolygons,
  destinationPoints, destinationPointsRow = 1,
  startDateAndTime = "2018-08-18 12:00:00", modes = "WALK, TRANSIT",
  maxWalkDistance = 1000, walkReluctance = 2, walkSpeed = 1.5,
  bikeSpeed = 5, minTransferTime = 1, maxTransfers = 5,
  wheelchair = F, arriveBy = F, durationCutoff = 60,
  waitingCutoff = 10, transferCutoff = 1,
  mapPolygonColours = "Blues", mapPolygonCatColours = c("#820e0e",
  "#407746"), mapZoom = "bb", mapPolygonLineWeight = 1,
  mapPolygonLineOpacity = 1, mapPolygonFillOpacity = 0.6,
  mapMarkerOpacity = 1, mapLegendOpacity = 0.5)
}
\arguments{
\item{output.dir}{The directory for the output files}

\item{otpcon}{The OTP router URL, see ?otpcon for details}

\item{originPoints}{The variable containing origin(s), see ?importLocationData for details}

\item{originPolygons}{The variable containing origin(s) polygon(s), see ?importGeojsonData for details}

\item{destinationPoints}{The variable containing destination(s) see ?importLocationData for details}

\item{destinationPointsRow}{The row of destinationPoints to be used, defaults to 1}

\item{startDateAndTime}{The start time and date, in 'YYYY-MM-DD HH:MM:SS' format}

\item{modes}{The mode of the journey, defaults to 'TRANSIT, WALK'}

\item{maxWalkDistance}{The maximum walking distance, in meters, defaults to 1000 m}

\item{walkReluctance}{The reluctance of walking-based routes, defaults to 2 (range 0 (lowest) - 20 (highest))}

\item{walkSpeed}{The walking soeed, in meters per second, defaults to 1.4 m/s}

\item{bikeSpeed}{The cycling speed, in meters per second, defaults to 4.3 m/s}

\item{minTransferTime}{The maximum transfer time, in minutes, defaults to 0 mins (no time specified)}

\item{maxTransfers}{The maximum number of transfers, defaults to 10}

\item{wheelchair}{If TRUE, uses on wheeelchair friendly stops, defaults to FALSE}

\item{arriveBy}{Selects whether journey starts at startDateandTime (FALSE) or finishes (TRUE), defaults to FALSE}

\item{durationCutoff}{For categorical maps, specify the duration time cutoff in minutes, defaults to 60 minutes}

\item{waitingCutoff}{For categorical maps, specify the waiting time cutoff in minutes, defaults to 10 minutes}

\item{transferCutoff}{For categorical maps, specify the number of transfers cutoff, defaults to 1}

\item{mapPolygonColours}{the color palette of the poylgon, defaults to 'Blue'}

\item{mapPolygonCatColours}{the color palette of the polygon for catergorical maps, defaults to c("#820e0e", "#407746")}

\item{mapZoom}{The zoom level of the map as an integer (e.g. 12), defaults to bounding box approach}

\item{mapPolygonLineWeight}{Specifies the weight of the polygon, defaults to 1 px}

\item{mapPolygonLineOpacity}{Specifies the opacity of the polygon line, defaults to 1 (solid)}

\item{mapPolygonFillOpacity}{Specifies the opacity of the polygon fill, defaults to 0.6}

\item{mapMarkerOpacity}{Specifies the opacity of the marker, defaults to 1 (solid)}

\item{mapLegendOpacity}{Specifies the opacity of the legend, defaults to 0.5}
}
\value{
Returns a series of maps (duration, wait time, transfers) to the output directory
}
\description{
Calculates a series of choropleth maps using multiple origins to a single destination.
}
\examples{
  choropleth(
    output.dir = 'C:\\Users\\User\\Documents',
    otpcon,
    originPoints,
    originPolygons,
    destinationPoints,
    startdDateAndTime = "2018-08-18 12:00:00"
  )
}
\author{
Michael Hodge
}
