% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{importGeojsonData}
\alias{importGeojsonData}
\title{Imports GeoJSON Data}
\usage{
importGeojsonData(src)
}
\arguments{
\item{src}{The source for the .GeoJSON file}
}
\value{
Large SpatialPolygonsDataFrame of origin polygons
}
\description{
Imports polygon data from a GeoJSON file
}
\author{
Michael Hodge
}
