% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{importLocationData}
\alias{importLocationData}
\title{Imports Location Data}
\usage{
importLocationData(src, idcol = "name", loncol = "lon",
  latcol = "lat", postcodecol = "postcode", deleteRow = T,
  autoComplete = F)
}
\arguments{
\item{src}{The source of the .CSV file}

\item{idcol}{The title of the header of the column containing unique identifier names, default is "name"}

\item{loncol}{The title of the header of the column containing the longitudinal values, default is "lon"}

\item{latcol}{The title of the header of the column containing the latitudinal values, default is "lat"}

\item{postcodecol}{The title of the header of the column containing the postcode values, default is "postcode"}

\item{deleteRow}{Specify whether to delete the row if no postcode is found, defaults to T}

\item{autoComplete}{Specify whether to use autocomplete if postcode is not valid}
}
\value{
R dataframe of location points
}
\description{
Imports location data from a comma separated file containing either latitude and longitude values, or postcode values
}
\author{
Michael Hodge
}
