% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isochrone-multi.R
\name{isochroneMulti}
\alias{isochroneMulti}
\title{Generates an isochrone map for multiple origins.}
\usage{
isochroneMulti(output.dir, otpcon, originPoints, destinationPoints,
  journeyReturn = F, startDateAndTime = format(Sys.time(),
  "\%Y-\%m-\%d \%H:\%M:\%S"), modes = "WALK, TRANSIT",
  maxWalkDistance = 1000, walkReluctance = 2, walkSpeed = 1.5,
  bikeSpeed = 5, minTransferTime = 1, maxTransfers = 5,
  wheelchair = F, arriveBy = F, isochroneCutOffMax = 90,
  isochroneCutOffMin = 10, isochroneCutOffStep = 10, infoPrint = T,
  mapOutput = F, geojsonOutput = F, failSafeSave = 100,
  mapZoom = "bb", mapPolygonLineWeight = 0,
  mapPolygonLineColor = "white", mapPolygonLineOpacity = 1,
  mapPolygonFillOpacity = 1, originMarker = T,
  originMarkerColor = "red", destinationMarkerSize = 3,
  destinationMarkerOpacity = 1, destinationMarkerStroke = T,
  destinationMarkerStrokeColor = "black",
  destinationMarkerStrokeWeight = 1, destinationMarkerColor = "white",
  mapLegendOpacity = 1, mapDarkMode = F)
}
\arguments{
\item{output.dir}{The directory for the output files}

\item{otpcon}{The OTP router URL, see ?otpcon for details}

\item{originPoints}{The variable containing origin(s), see ?importLocationData for details}

\item{destinationPoints}{The variable containing destination(s) see ?importLocationData for details}

\item{journeyReturn}{Specifies whether the journey should be calculated as a return or not (default is FALSE)}

\item{startDateAndTime}{The start time and date, in 'YYYY-MM-DD HH:MM:SS' format, default is currrent date and time}

\item{modes}{The mode of the journey, defaults to 'TRANSIT, WALK'}

\item{maxWalkDistance}{The maximum walking distance, in meters, defaults to 1000 m}

\item{walkReluctance}{The reluctance of walking-based routes, defaults to 2 (range 0 (lowest) - 20 (highest))}

\item{walkSpeed}{The walking soeed, in meters per second, defaults to 1.4 m/s}

\item{bikeSpeed}{The cycling speed, in meters per second, defaults to 4.3 m/s}

\item{minTransferTime}{The maximum transfer time, in minutes, defaults to 0 mins (no time specified)}

\item{maxTransfers}{The maximum number of transfers, defaults to 10}

\item{wheelchair}{If TRUE, uses on wheeelchair friendly stops, defaults to FALSE}

\item{arriveBy}{Selects whether journey starts at startDateandTime (FALSE) or finishes (TRUE), defaults to FALSE}

\item{isochroneCutOffMax}{Provide the maximum cutoff time for the isochrone, defaults 90}

\item{isochroneCutOffMin}{Provide the minimum cutoff time for the isochrone, defaults 10}

\item{isochroneCutOffStep}{Provide the cutoff time step for the isochrone, 0 denotes no step is required (returns isochroneCutOffMax only), defaults 10}

\item{infoPrint}{Specifies whether you want some information printed to the console or not, default is TRUE}

\item{mapOutput}{Specifies whether you want to output a map, defaults to FALSE}

\item{geojsonOutput}{Specifies whether you want to output a GeoJSON file, defaults to FALSE}

\item{failSafeSave}{Specify the failsafe save number for large datasets, default is 100}

\item{mapZoom}{The zoom level of the map as an integer (e.g. 12), defaults to bounding box approach}

\item{mapPolygonLineWeight}{Specifies the weight of the polygon, defaults to 0 px}

\item{mapPolygonLineColor}{Specifies the color of the polygon, defaults to 'white'}

\item{mapPolygonLineOpacity}{Specifies the opacity of the polygon line, defaults to 1 (solid)}

\item{mapPolygonFillOpacity}{Specifies the opacity of the polygon fill, defaults to 1}

\item{originMarker}{Specifies if you want to output the origin markers to the map (default is True)}

\item{originMarkerColor}{Specifies the colour of the origin marker if it is within a isochrone (default is 'red')}

\item{destinationMarkerSize}{Specifies the destination marker(s) size (default is 3)}

\item{destinationMarkerOpacity}{Specifies the opacity of destination marker(s)if it is within a isochrone (default is 1, solid)}

\item{destinationMarkerStroke}{Specifies whether a destination marker(s) stroke is used (default is T)}

\item{destinationMarkerStrokeColor}{Specifies the stroke color for the destination marker(s) (default is 'black')}

\item{destinationMarkerStrokeWeight}{Specifies the marker stroke weight for the destination marker(s) (default is 1)}

\item{destinationMarkerColor}{Specifies the colour of destination marker(s) if it is not within a isochrone (default is 'white')}

\item{mapLegendOpacity}{Specifies the opacity of the legend, defaults to 1}

\item{mapDarkMode}{Specifies if you want to use the dark leaflet map colour (default is FALSE)}
}
\value{
Saves journey details as CSV to output directory (optional: a map in PNG and HTML formats, the polygons as a GeoJSON)
}
\description{
Generates an isochrone map from multiple origins and checks whether destinations
fall within isochrones, and if so, at what cutoff time amount.
A comma separated value file of journey times for each origin and destination is saved in the output folder.
A map of the journey can also be saved as a .png image and .html file.
The polygons can also be saved as a .GeoJSON file.
}
\author{
Michael Hodge
}
