% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isochrone-multi-intersect.R
\name{isochroneMultiIntersect}
\alias{isochroneMultiIntersect}
\title{Isochrone map for multiple origins}
\usage{
isochroneMultiIntersect(output.dir, otpcon, originPoints, destinationPoints,
  startDateAndTime = "2018-08-18 12:00:00", modes = "WALK, TRANSIT",
  maxWalkDistance = 1000, walkReluctance = 2, walkSpeed = 1.5,
  bikeSpeed = 5, minTransferTime = 1, maxTransfers = 5, wheelchair = F,
  arriveBy = F, isochroneCutOffs = 60, palColorMarker = "Greys",
  palColorPolygon = "Purple", mapZoom = 12)
}
\arguments{
\item{output.dir}{The directory for the output files}

\item{otpcon}{OTP router URL}

\item{originPoints}{The variable containing origin(s), see ?importLocationData}

\item{destinationPoints}{The variable containing destination(s) see ?importLocationData}

\item{startDateAndTime}{in 'YYYY-MM-DD HH:MM:SS' format}

\item{modes}{defaults to 'TRANSIT, WALK'}

\item{maxWalkDistance}{in meters, defaults to 1000}

\item{walkReluctance}{defaults to 2 (range 0 - 20)}

\item{walkSpeed}{in m/s, defaults to 1.4}

\item{bikeSpeed}{in m/s, defaults to 4.3}

\item{minTransferTime}{in minutes, defaults to 1}

\item{maxTransfers}{defaults to 10}

\item{wheelchair}{defaults to FALSE}

\item{arriveBy}{defaults to FALSE}

\item{isochroneCutOffs}{in minutes, defaults to 60}

\item{palColorPolygon}{the color palette of the poygon, defaults to 'Purple'}

\item{mapZoom}{defaults to 12}

\item{palColorMarkers}{the color palette of the markers, defaults to 'Greys'}
}
\value{
Saves map as a png and geojson of the intersection area to output directory
}
\description{
Calculates an isochrone map for multiple origins and finds the intersection.
}
\examples{
  isochroneMultiIntersect(
    output.dir = 'C:\\Users\\User\\Documents',
    otpcon,
    originPoints,
    destinationPoints,
    startDateAndTime = "2018-08-18 12:00:00"
  )
}
\author{
Michael Hodge
}
