% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isochrone-time.R
\name{isochroneTime}
\alias{isochroneTime}
\title{Generates an isochrone map from a single origin between a start and end
time and date.}
\usage{
isochroneTime(output.dir, otpcon, originPoints, originPointsRow = 1,
  destinationPoints, destinationPointsRow = 1,
  startDateAndTime = format(Sys.time(), "\%Y-\%m-\%d \%H:\%M:\%S"),
  endDateAndTime = format(Sys.time() + 1 * 60 * 60,
  "\%Y-\%m-\%d \%H:\%M:\%S"), timeIncrease = 60,
  modes = "WALK, TRANSIT", maxWalkDistance = 1000,
  walkReluctance = 2, walkSpeed = 1.5, bikeSpeed = 5,
  minTransferTime = 1, maxTransfers = 5, wheelchair = F,
  arriveBy = F, isochroneCutOffMax = 90, isochroneCutOffMin = 10,
  isochroneCutOffStep = 10, infoPrint = T, gifOutput = F,
  mapOutput = F, geojsonOutput = F, mapPolygonColours = "Blues",
  mapZoom = "bb", mapPolygonLineWeight = 0,
  mapPolygonLineColor = "white", mapPolygonLineOpacity = 1,
  mapPolygonFillOpacity = 1, originMarker = T,
  originMarkerColor = "red", destinationMarkerSize = 3,
  destinationMarkerOutOpacity = 1, destinationMarkerInOpacity = 1,
  destinationMarkerStroke = T, destinationMarkerStrokeColor = "black",
  destinationMarkerStrokeWeight = 1,
  destinationMarkerInColor = "white",
  destinationMarkerOutColor = "grey", mapLegendOpacity = 1,
  mapDarkMode = F)
}
\arguments{
\item{output.dir}{The directory for the output files}

\item{otpcon}{The OTP router URL, see ?otpcon for details}

\item{originPoints}{The variable containing origin(s), see ?importLocationData for details}

\item{originPointsRow}{The row of originPoints to be used, defaults to 1}

\item{destinationPoints}{The variable containing destination(s) see ?importLocationData for details}

\item{startDateAndTime}{The start time and date, in 'YYYY-MM-DD HH:MM:SS' format, default is currrent date and time}

\item{endDateAndTime}{The end time and date, in 'YYYY-MM-DD HH:MM:SS' format, default is an hour after the currrent date and time}

\item{timeIncrease}{The time increase in minutes, default 60}

\item{modes}{The mode of the journey, defaults to 'TRANSIT, WALK'}

\item{maxWalkDistance}{The maximum walking distance, in meters, defaults to 1000 m}

\item{walkReluctance}{The reluctance of walking-based routes, defaults to 2 (range 0 (lowest) - 20 (highest))}

\item{walkSpeed}{The walking soeed, in meters per second, defaults to 1.4 m/s}

\item{bikeSpeed}{The cycling speed, in meters per second, defaults to 4.3 m/s}

\item{minTransferTime}{The maximum transfer time, in minutes, defaults to 0 mins (no time specified)}

\item{maxTransfers}{The maximum number of transfers, defaults to 10}

\item{wheelchair}{If TRUE, uses on wheeelchair friendly stops, defaults to FALSE}

\item{arriveBy}{Selects whether journey starts at startDateandTime (FALSE) or finishes (TRUE), defaults to FALSE}

\item{isochroneCutOffMax}{Provide the maximum cutoff time for the isochrone, defaults 90}

\item{isochroneCutOffMin}{Provide the minimum cutoff time for the isochrone, defaults 10}

\item{isochroneCutOffStep}{Provide the cutoff time step for the isochrone, defaults 10}

\item{infoPrint}{Specifies whether you want some information printed to the console or not, default is TRUE}

\item{gifOutput}{Specifies whether you want to output a gif, defaults to FALSE}

\item{mapOutput}{Specifies whether you want to output the maps, defaults to FALSE}

\item{geojsonOutput}{Specifies whether you want to output a GeoJSON file, defaults to FALSE}

\item{mapPolygonColours}{The color palette of the map, defaults to 'Blues'}

\item{mapZoom}{The zoom level of the map as an integer (e.g. 12), defaults to bounding box approach}

\item{mapPolygonLineWeight}{Specifies the weight of the polygon, defaults to 0 px}

\item{mapPolygonLineColor}{Specifies the color of the polygon, defaults to 'white'}

\item{mapPolygonLineOpacity}{Specifies the opacity of the polygon line, defaults to 1 (solid)}

\item{mapPolygonFillOpacity}{Specifies the opacity of the polygon fill, defaults to 1}

\item{originMarker}{Specifies if you want to output the origin markers to the map (default is True)}

\item{originMarkerColor}{Specifies the colour of the origin marker if it is within a isochrone (default is 'red')}

\item{destinationMarkerSize}{Specifies the destination marker(s) size (default is 3)}

\item{destinationMarkerOutOpacity}{Specifies the opacity of destination marker(s)if it is not within a isochrone (default is 1, solid)}

\item{destinationMarkerInOpacity}{Specifies the opacity of destination marker(s)if it is within a isochrone (default is 1, solid)}

\item{destinationMarkerStroke}{Specifies whether a destination marker(s) stroke is used (default is T)}

\item{destinationMarkerStrokeColor}{Specifies the stroke color for the destination marker(s) (default is 'black')}

\item{destinationMarkerStrokeWeight}{Specifies the marker stroke weight for the destination marker(s) (default is 1)}

\item{destinationMarkerInColor}{Specifies the colour of destination marker(s)if it is within a isochrone (default is 'white')}

\item{destinationMarkerOutColor}{Specifies the colour of destination marker(s) if it is not within a isochrone (default is 'grey')}

\item{mapLegendOpacity}{Specifies the opacity of the legend, defaults to 1}

\item{mapDarkMode}{Specifies if you want to use the dark leaflet map colour (default is FALSE)}
}
\value{
Saves journey details as comma separated value file to output directory. An animated map in .gif format may also be saved.
}
\description{
Generates isochrone polygons from a single origin between a start and end
time and date, and checks whether destinations fall within isochrone,
and if so, at what cutoff time amount.
A comma separated value file of journey times for each destination (and time) is saved in the output folder.
An animated map of the journey can also be saved as a .gif image.
}
\author{
Michael Hodge
}
