% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point-to-point.R
\name{pointToPoint}
\alias{pointToPoint}
\title{Calculates the journey for a single origin and destination.}
\usage{
pointToPoint(output.dir, otpcon, originPoints, originPointsRow = 1,
  destinationPoints, destinationPointsRow = 1,
  startDateAndTime = format(Sys.time(), "\%Y-\%m-\%d \%H:\%M:\%S"),
  modes = "WALK, TRANSIT", maxWalkDistance = 1000,
  walkReluctance = 2, walkSpeed = 1.4, bikeSpeed = 4.3,
  minTransferTime = 1, maxTransfers = 10, wheelchair = F,
  arriveBy = F, preWaitTime = 15, estimateCost = F,
  busTicketPrice = 3, busTicketPriceMax = 12,
  trainTicketPriceKm = 0.12, trainTicketPriceMin = 3, infoPrint = T,
  mapOutput = F, geojsonOutput = F, mapPolylineColours = list(TRANSIT
  = "#000000", WALK = "#A14296", BUS = "#48C1B1", RAIL = "#4D7BC5", CAR =
  "#E825D6", BICYCLE = "#4AA6C3"), mapZoom = "bb",
  mapPolylineWeight = 5, mapPolylineOpacity = 1,
  mapMarkerStrokeColor = "black", mapMarkerStrokeWeight = 1,
  mapMarkerOpacity = 1, mapLegendOpacity = 1, mapDarkMode = F)
}
\arguments{
\item{output.dir}{The directory for the output files}

\item{otpcon}{The OTP router URL, see ?otpcon for details}

\item{originPoints}{The variable containing origin(s), see ?importLocationData for details}

\item{originPointsRow}{The row of originPoints to be used, defaults to 1}

\item{destinationPoints}{The variable containing destination(s) see ?importLocationData for details}

\item{destinationPointsRow}{The row of destinationPoints to be used, defaults to 1}

\item{startDateAndTime}{The start time and date, in 'YYYY-MM-DD HH:MM:SS' format, default is currrent date and time}

\item{modes}{The mode of the journey, defaults to 'TRANSIT, WALK'}

\item{maxWalkDistance}{The maximum walking distance, in meters, defaults to 1000 m}

\item{walkReluctance}{The reluctance of walking-based routes, defaults to 2 (range 0 (lowest) - 20 (highest))}

\item{walkSpeed}{The walking soeed, in meters per second, defaults to 1.4 m/s}

\item{bikeSpeed}{The cycling speed, in meters per second, defaults to 4.3 m/s}

\item{minTransferTime}{The maximum transfer time, in minutes, defaults to 0 mins (no time specified)}

\item{maxTransfers}{The maximum number of transfers, defaults to 10}

\item{wheelchair}{If TRUE, uses on wheeelchair friendly stops, defaults to FALSE}

\item{arriveBy}{Selects whether journey starts at startDateandTime (FALSE) or finishes (TRUE), defaults to FALSE}

\item{preWaitTime}{The maximum waiting time before a journey cannot be found, in minutes, defaults to 15 mins}

\item{estimateCost}{Specify whether to estimate costs of journey or not (default is False)}

\item{busTicketPrice}{Specifiy the cost of a bus journey (default is 3 GPB)}

\item{busTicketPriceMax}{Specifiy the maximum cost of a bus journey (default is 12 GPB)}

\item{trainTicketPriceKm}{Specifiy the cost of a train journey per km (default is 0.12 GPB per km)}

\item{trainTicketPriceMin}{Specifiy the minimum cost of a train journey (default is 3 GBP)}

\item{infoPrint}{Specifies whether you want some information printed to the console or not, default is TRUE}

\item{mapOutput}{Specifies whether you want to output a map, defaults to FALSE}

\item{geojsonOutput}{Specifies whether you want to output the polylines as a geojson, defaults to FALSE}

\item{mapPolylineColours}{A list defining the colours to assign to each mode of transport.}

\item{mapZoom}{The zoom level of the map as an integer (e.g. 12), defaults to bounding box approach}

\item{mapPolylineWeight}{Specifies the weight of the polyline, defaults to 5 px}

\item{mapPolylineOpacity}{Specifies the opacity of the polyline, defaults to 1 (solid)}

\item{mapMarkerStrokeColor}{Specifies the outline color of the marker, defaults to black}

\item{mapMarkerStrokeWeight}{Specifies the stroke weight of the marker, defaults to 1 (solid)}

\item{mapMarkerOpacity}{Specifies the opacity of the marker, defaults to 1 (solid)}

\item{mapLegendOpacity}{Specifies the opacity of the legend, defaults to 1 (solid)}

\item{mapDarkMode}{Specifies if you want to use the dark leaflet map colour (default is FALSE)}
}
\value{
Saves journey details as comma separated value file to output directory. A map in .png and .html formats may also be saved)
}
\description{
Calculates the journey time and details between a single origin and destination. 
A comma separated value file of journey details is saved in the specified output folder.
A map of the journey can also be saved as a .png image and .html file.
}
\author{
Michael Hodge
}
