% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point-to-point-time.R
\name{pointToPointTime}
\alias{pointToPointTime}
\title{Calculates the journey between for a single origin and destination between a
start and end time and date.}
\usage{
pointToPointTime(output.dir, otpcon, originPoints, originPointsRow = 1,
  destinationPoints, destinationPointsRow = 1,
  startDateAndTime = "2018-08-18 12:00:00",
  endDateAndTime = "2018-08-18 15:00:00", timeIncrease = 60,
  modes = "WALK, TRANSIT", maxWalkDistance = 1000, walkReluctance = 2,
  walkSpeed = 1.5, bikeSpeed = 5, minTransferTime = 1, maxTransfers = 5,
  wheelchair = F, arriveBy = F, preWaitTime = 60,
  transportColours = list(TRANSIT = "#000000", WALK = "#A14296", BUS =
  "#48C1B1", RAIL = "#4D7BC5", CAR = "#8D4084", BICYCLE = "#4AA6C3"),
  mapZoom = 12)
}
\arguments{
\item{output.dir}{The directory for the output files}

\item{otpcon}{OTP router URL}

\item{originPoints}{The variable containing origin(s), see ?importLocationData}

\item{originPointsRow}{The row of originPoints to be used, defaults to 1}

\item{destinationPoints}{The variable containing destination(s) see ?importLocationData}

\item{destinationPointsRow}{The row of destinationPoints to be used, defaults to 1}

\item{startDateAndTime}{in 'YYYY-MM-DD HH:MM:SS' format}

\item{endDateAndTime}{in 'YYYY-MM-DD HH:MM:SS' format}

\item{timeIncrease}{in minutes, default 60}

\item{modes}{defaults to 'TRANSIT, WALK'}

\item{maxWalkDistance}{in meters, defaults to 1000}

\item{walkReluctance}{defaults to 2 (range 0 - 20)}

\item{walkSpeed}{in m/s, defaults to 1.4}

\item{bikeSpeed}{in m/s, defaults to 4.3}

\item{minTransferTime}{in minutes, defaults to 1}

\item{maxTransfers}{defaults to 10}

\item{wheelchair}{defaults to FALSE}

\item{arriveBy}{defaults to FALSE}

\item{preWaitTime}{in minutes, defaults to 60}

\item{transportColours}{A list defining the colours to assign to each mode of transport.}

\item{mapZoom}{defaults to 12}
}
\value{
Saves an animated map as a gif and journey details as CSV to output directory
}
\description{
Calculates the journey time and details between a single origin and destination between a start
and end time and date. Outputs (and saves) a animated gif map of the polyline between origin and 
destination, as well as a CSV file of journey details.
}
\examples{
  pointToPointTime(
    output.dir = 'C:\\Users\\User\\Documents',
    otpcon,
    originPoints,
    destinationPoints,
    startDateAndTime = "2018-08-18 12:00:00",
    endDateAndTime = "2018-08-18 13:00:00",
    timeIncrease = 60
  )
}
\author{
Michael Hodge
}
